MASTER: Initial commit
Essentially contains the default WordPress package, well as of
December 3rd, 2014 that is -- with a couple theme deletions and
a bunch of installation specific files to ignore -- in the
'access' directory. And an 'index.php' file sitting at the root
pointing to 'wp-blog-header.php' allowing WordPress to function
from the root. Also includes repository files, a license (GNU
GPL v2.0) and a README. Note: the 'wp-content' directory is
renamed 'abled-content'.