MASTER: Reformat 'primary-sidebar' in style and behaviour
A somewhat custom solution. In the sense it's a solution for a
problem that presents itself as a result of how I use both the
archives and categories lists in the 'primary-sidebar'. Now the
overall style of the 'primary-sidebar' matches 'content-sidebar'
in looks at wide viewports. But at narrow viewport width's it,
when the 'primary-sidebar' is forced below the 'main-content'
that is, hides the lists and shows the same information through
dropdown menus to save some space.