MASTER: Provide user more information, in context
At narrow viewport width's, once the 'mobile menu' shows, a user
is also given the word 'menu' beside the 'hamburger icon', to
provide that much more information as to what the icon means, and
hopefully to better indicate what it does. And the 'search toggle'
remains in the same place regardless of viewport width.